Add redux to monthly
