+++ src/views/categorize.jsx
